+++ changelog.md
@@ -1,7 +1,48 @@
 Changelog
 =========
 
+1.0.6
+-----
+  - Fix: Allow directories inside the base directory
+  - Fix: Parser wasn't handling the values of @css lines correctly
+  - Improvement: Allow explicit closing of @media blocks using "@media none"
+  - Improvement: Fail silently when version.txt goes missing
+  - Improvement: Display an error when a plugin that doesn't exist is called
+  - Improvement: Display an error when a file outside the base directory is being accessed
+
+
+1.0.5
+-----
+  - Fix: Prevent directory traversal by only allowing turbine to access files in the css_base_dir
+  - Fix: Made background gradient plugin more tolerant to whitespace
+
+
+1.0.4
+-----
+  - Fix: Background gradient sometimes didn't work in Opera
+  - Fix: Multiple filters in IE8 didn't work
+
+
+1.0.3
+-----
+  - Fix: Inheritance from multiple instances of the same selector failed
+  - Improvement: Better debugging output when dealing with inheritance
+
+
+1.0.2
+-----
+  - Fix: Bugfix plugin inserted unnecessary behavior values
+  - Fix: Wrong paths for htc files in bugfix plugin
+  - Fix: Removed non-existing IE6 plugin from legacy meta plugin
+
+
+1.0.1
+-----
+  - Fix: IE7 didn't recieve images as MHTML
+  - Fix: Loader plugin preserved original file paths for included modules
+  - Fix: Box sizing wasn't part of CSS3 meta plugin
+
+
 1.0.0
 -----
-
   - Initial Version